requirements.txt
django>=1.4

# tastypie
django-tastypie

jsonfield